completed user-can-view-existing-characters
